Criando rota para alteração de usuários e veículos. Realizando as tratativas de exceção na criação de veículos para cada tipo de variavel

+++ api/routes/veiculos/delecaoVeiculos.js
@@ -2,13 +2,22 @@ const express = require("express");
 const DelecaoVeiculos = express();
 const BancoVeiculos = require("../../../database/veiculos/veiculos");
 
-DelecaoVeiculos.delete("/DelecaoVeiculos/:id", (req, res) => {
-  BancoVeiculos.destroy({
-    where: {
-      id: req.params.id,
-    },
-  });
-  res.send("Veiculo deletado com sucesso");
+DelecaoVeiculos.delete("/delecao-Veiculos/:id", (req, res) => {
+  try {
+    if(req.params.id){
+      BancoVeiculos.destroy({
+        where: {
+          id: req.params.id,
+        },
+      });
+      res.send("Veiculo deletado com sucesso");
+    } else {
+      res.send("Veículo não existe");
+    }
+
+  } catch (error) {
+    res.send("Erro").statusCode(404);
+  }
 });
 
 module.exports = DelecaoVeiculos;
